# docs/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "description" : "Incidents, service requests, workflow, assignment, comments and audit for IT service management. Errors are RFC 9457 problem details (`ApiProblem`); clients should branch on `code`.",
    "title" : "Enterprise Service Management API",
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "/",
    "description" : "Same origin as the web app"
  } ],
  "security" : [ {
    "bearer-jwt" : [ ]
  } ],
  "paths" : {
    "/api/admin/categories" : {
      "get" : {
        "operationId" : "listAllCategories",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/CategoryResponse"
                  }
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Error: see `code` (validation 400, forbidden 403, not found 404, conflict 409, rule violation 422)"
          }
        },
        "summary" : "All categories, including inactive ones",
        "tags" : [ "Categories" ]
      },
      "post" : {
        "operationId" : "createCategory",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateCategoryRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CategoryResponse"
                }
              }
            },
            "description" : "Created"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Error: see `code` (validation 400, forbidden 403, not found 404, conflict 409, rule violation 422)"
          }
        },
        "summary" : "Create a category or subcategory",
        "tags" : [ "Categories" ]
      }
    },
    "/api/admin/categories/{id}" : {
      "patch" : {
        "operationId" : "updateCategory",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateCategoryRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CategoryResponse"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Error: see `code` (validation 400, forbidden 403, not found 404, conflict 409, rule violation 422)"
          }
        },
        "summary" : "Rename, re-route or (de)activate a category (requires current version)",
        "tags" : [ "Categories" ]
      }
    },
...
    "/api/categories" : {
      "get" : {
        "operationId" : "listCategories",
        "parameters" : [ {
          "in" : "query",
          "name" : "type",
          "required" : true,
          "schema" : {
            "type" : "string",
            "enum" : [ "INCIDENT", "SERVICE_REQUEST" ]
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/CategoryResponse"
                  }
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Error: see `code` (validation 400, forbidden 403, not found 404, conflict 409, rule violation 422)"
          }
        },
        "summary" : "Active categories (with subcategories) usable for a ticket type",
        "tags" : [ "Categories" ]
      }
    },
...
    "/api/tickets/{id}/assignment" : {
      "put" : {
        "operationId" : "assignTicket",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AssignmentRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AssignmentResponse"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Error: see `code` (validation 400, forbidden 403, not found 404, conflict 409, rule violation 422)"
          }
        },
        "summary" : "Take, release, assign, reassign or transfer a ticket (requires current version)",
        "tags" : [ "Ticket assignment" ]
      }
    },
    "/api/tickets/{id}/comments" : {
      "get" : {
        "operationId" : "listComments",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/CommentResponse"
                  }
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Error: see `code` (validation 400, forbidden 403, not found 404, conflict 409, rule violation 422)"
          }
        },
        "summary" : "The ticket's conversation, oldest first. Internal notes only for support staff",
        "tags" : [ "Ticket comments" ]
      },
      "post" : {
        "operationId" : "addComment",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AddCommentRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CommentResponse"
                }
              }
            },
            "description" : "Created"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Error: see `code` (validation 400, forbidden 403, not found 404, conflict 409, rule violation 422)"
          }
        },
        "summary" : "Add a public comment or (support staff) an internal note",
        "tags" : [ "Ticket comments" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "AddCommentRequest" : {
        "type" : "object",
        "properties" : {
          "body" : {
            "type" : "string",
            "maxLength" : 10000,
            "minLength" : 0
          },
          "visibility" : {
            "type" : "string",
            "enum" : [ "PUBLIC", "INTERNAL" ]
          }
        },
        "required" : [ "body", "visibility" ]
      },
      "ApiProblem" : {
        "type" : "object",
        "description" : "RFC 9457 problem details, returned for every error",
        "properties" : {
          "code" : {
            "type" : "string",
            "description" : "Stable machine-readable code: branch on this, never on detail",
            "enum" : [ "VALIDATION_FAILED", "MALFORMED_REQUEST", "AUTHENTICATION_REQUIRED", "ACCESS_DENIED", "RESOURCE_NOT_FOUND", "METHOD_NOT_ALLOWED", "UNSUPPORTED_MEDIA_TYPE", "CONCURRENT_MODIFICATION", "INVALID_SORT_PROPERTY", "INVALID_CREDENTIALS", "REFRESH_TOKEN_INVALID", "INVALID_ORIGIN", "CURRENT_PASSWORD_INCORRECT", "EMAIL_ALREADY_EXISTS", "LAST_ADMIN_REQUIRED", "TEAM_NAME_ALREADY_EXISTS", "TEAM_INACTIVE", "USER_NOT_ELIGIBLE_FOR_TEAM", "CATEGORY_CODE_ALREADY_EXISTS", "CATEGORY_TEAM_REQUIRED", "INVALID_CATEGORY", "INVALID_STATUS_TRANSITION", "TRANSITION_NOT_PERMITTED", "TRANSITION_REQUIREMENT_MISSING", "TICKET_NOT_ASSIGNED", "ASSIGNMENT_NOT_PERMITTED", "ASSIGNEE_NOT_IN_TEAM", "TICKET_NOT_ASSIGNABLE", "MEMBER_HAS_OPEN_TICKETS", "COMMENT_NOT_PERMITTED", "TICKET_CLOSED", "REQUEST_FAILED", "INTERNAL_ERROR" ]
          },
          "correlationId" : {
            "type" : [ "string", "null" ],
            "description" : "Matches the X-Request-Id response header and server log lines"
          },
          "detail" : {
            "type" : [ "string", "null" ],
            "description" : "Human-readable explanation; never contains internal details"
          },
          "fieldErrors" : {
            "type" : [ "array", "null" ],
            "description" : "Present for VALIDATION_FAILED",
            "items" : {
              "$ref" : "#/components/schemas/FieldErrorDetail"
            }
          },
          "instance" : {
            "type" : [ "string", "null" ],
            "description" : "Request path",
            "example" : "/api/tickets/0199b2c4-6a1e-7c3e-9f00-1a2b3c4d5e6f/transitions"
          },
          "status" : {
            "type" : "integer",
            "format" : "int32",
            "example" : 409
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          },
          "title" : {
            "type" : "string",
            "example" : "Conflict"
          },
          "type" : {
            "type" : "string",
            "description" : "Problem type URI",
            "example" : "about:blank"
          }
        },
        "required" : [ "code", "correlationId", "detail", "fieldErrors", "instance", "status", "timestamp", "title", "type" ]
      },
      "AssignmentRequest" : {
        "type" : "object",
        "properties" : {
          "assigneeId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "teamId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "version" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
        "required" : [ "teamId", "version" ]
      },
      "AssignmentResponse" : {
        "type" : "object",
        "properties" : {
          "assignedTeam" : {
            "$ref" : "#/components/schemas/NamedRef"
          },
          "assignee" : {
            "oneOf" : [ {
              "$ref" : "#/components/schemas/NamedRef"
            }, {
              "type" : "null"
            } ]
          },
          "reference" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "ticketId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "version" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
        "required" : [ "assignedTeam", "assignee", "reference", "status", "ticketId", "version" ]
      },
...
      "CategoryResponse" : {
        "type" : "object",
        "properties" : {
          "active" : {
            "type" : "boolean"
          },
          "appliesTo" : {
            "type" : "string",
            "enum" : [ "INCIDENT", "SERVICE_REQUEST", "ANY" ]
          },
          "code" : {
            "type" : "string"
          },
          "defaultTeamId" : {
            "type" : [ "string", "null" ],
            "format" : "uuid"
          },
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "parentId" : {
            "type" : [ "string", "null" ],
            "format" : "uuid"
          },
          "subcategories" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CategoryResponse"
            }
          },
          "version" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
        "required" : [ "active", "appliesTo", "code", "defaultTeamId", "id", "name", "parentId", "subcategories", "version" ]
      },
...
      "CommentResponse" : {
        "type" : "object",
        "properties" : {
          "author" : {
            "$ref" : "#/components/schemas/NamedRef"
          },
          "body" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "relatedStatus" : {
            "type" : [ "string", "null" ]
          },
          "visibility" : {
            "type" : "string",
            "enum" : [ "PUBLIC", "INTERNAL" ]
          }
        },
        "required" : [ "author", "body", "createdAt", "id", "relatedStatus", "visibility" ]
      },
      "CreateCategoryRequest" : {
        "type" : "object",
        "properties" : {
          "appliesTo" : {
            "type" : "string",
            "enum" : [ "INCIDENT", "SERVICE_REQUEST", "ANY" ]
          },
          "code" : {
            "type" : "string",
            "minLength" : 1,
            "pattern" : "^[A-Z][A-Z0-9_]{1,49}$"
          },
          "defaultTeamId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string",
            "maxLength" : 100,
            "minLength" : 0
          },
          "parentId" : {
            "type" : "string",
            "format" : "uuid"
          }
        },
        "required" : [ "appliesTo", "code", "name" ]
      },
...
      "FieldErrorDetail" : {
        "type" : "object",
        "properties" : {
          "field" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          }
        },
        "required" : [ "field", "message" ]
      },
...
      "NamedRef" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          }
        },
        "required" : [ "id", "name" ]
      },
...
      "UpdateCategoryRequest" : {
        "type" : "object",
        "properties" : {
          "active" : {
            "type" : "boolean"
          },
          "defaultTeamId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string",
            "maxLength" : 100,
            "minLength" : 0,
            "pattern" : ".*\\S.*"
          },
          "version" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
        "required" : [ "version" ]
      },
...
    },
    "securitySchemes" : {
      "bearer-jwt" : {
        "bearerFormat" : "JWT",
        "description" : "Access token from POST /api/auth/login",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  }
}